// packages/core/src/input/enums/InputType.ts
/**
 * Several known input types and other types to be extended.
 */
export enum InputType {
  Pointer = 0,
  Wheel = 1,
  Keyboard = 2,
  XInput1 = 3,
  XInput2 = 4,
  XInput3 = 5,
  XInput4 = 6,
  XInput5 = 7,
  XInput6 = 8,
  XInput7 = 9,
  XInput8 = 10,
  XInput9 = 11,
  XInput10 = 12,
  length = 13
}